# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gochive apis",
        "title": "gochive",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/amingze/gochive"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/amingze/gochive/blob/master/LICENSE"
        },
        "version": "0.0.1"
    },
    "basePath": "/api/",
    "paths": {
...
        "/system/account": {
            "put": {
                "description": "初始化管理员账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "初始化管理员账号",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "roles",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "storage_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "ticket",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/system/database": {
            "put": {
                "description": "初始化数据库链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "初始化数据库",
                "parameters": [
                    {
                        "type": "string",
                        "name": "driver",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "dsn",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/system/matter-path-envs": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "admin"
                        ]
                    }
                ],
                "description": "环境变量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "环境变量",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/system/options/{name}": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "admin"
                        ]
                    }
                ],
                "description": "查找设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "查找设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "参数",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OAuth2Application": [
                            "admin"
                        ]
                    }
                ],
                "description": "更新设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "更新设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
        "/tokens": {
            "post": {
                "description": "用于账户登录和申请密码重置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "登录/密码重置",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bind.BodyToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "用户状态登出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bind.BodyToken": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {
                "type": "any"
            }
        },
...
        "httputil.JSONResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "/api/tokens",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "matter": " Grants matter access and write"
            }
        }
    }
}